added basic style in BottomTab and its Screen components
