+++ styles/AppStyles.js
@@ -1,0 +1,4 @@
+// General colors that used in the application
+export default {
+    mainColor: '#4E397E', 
+};
